requirements.txt
flask>=3.0
anthropic>=0.40
python-dotenv>=1.0
python-chess>=1.999
